maternal-health-risk/requirements.txt
sagemaker==2.*

pandas
xgboost
scikit-learn
joblib
numpy
python-dotenv

boto3
